Deep copy context when generating full context
