+++ tests/unit/structures/test_pipeline.py
@@ -411,48 +411,3 @@ class TestPipeline:
         pipeline.run()
         assert len(pipeline.task_outputs) == 1
         assert pipeline.task_outputs[task.id] == task.output
-
-    def test_to_dict(self):
-        task = PromptTask("test")
-        pipeline = Pipeline()
-        pipeline + [task]
-        expected_pipeline_dict = {
-            "type": pipeline.type,
-            "id": pipeline.id,
-            "tasks": [
-                {
-                    "type": pipeline.tasks[0].type,
-                    "id": pipeline.tasks[0].id,
-                    "state": str(pipeline.tasks[0].state),
-                    "parent_ids": pipeline.tasks[0].parent_ids,
-                    "child_ids": pipeline.tasks[0].child_ids,
-                    "max_meta_memory_entries": pipeline.tasks[0].max_meta_memory_entries,
-                    "context": pipeline.tasks[0].context,
-                }
-            ],
-            "conversation_memory": {
-                "type": pipeline.conversation_memory.type,
-                "runs": pipeline.conversation_memory.runs,
-                "meta": pipeline.conversation_memory.meta,
-                "max_runs": pipeline.conversation_memory.max_runs,
-            },
-            "conversation_memory_strategy": pipeline.conversation_memory_strategy,
-            "fail_fast": pipeline.fail_fast,
-        }
-        assert pipeline.to_dict() == expected_pipeline_dict
-
-    def test_from_dict(self):
-        task = PromptTask("test")
-        pipeline = Pipeline(tasks=[task])
-
-        serialized_pipeline = pipeline.to_dict()
-        assert isinstance(serialized_pipeline, dict)
-
-        deserialized_pipeline = Pipeline.from_dict(serialized_pipeline)
-        assert isinstance(deserialized_pipeline, Pipeline)
-
-        assert deserialized_pipeline.task_outputs[task.id] is None
-        deserialized_pipeline.run()
-
-        assert len(deserialized_pipeline.task_outputs) == 1
-        assert deserialized_pipeline.task_outputs[task.id].value == "mock output"
